feat(layer): create layer_manager to handle all layer related stuff
