+++ antarest/study/business/area_management.py
@@ -14,14 +14,13 @@ import logging
 import re
 from typing import Any, Dict, List, Mapping, Optional, Sequence
 
-from antarest.core.exceptions import ConfigFileNotFound, DuplicateAreaName, LayerNotAllowedToBeDeleted, LayerNotFound
+from antarest.core.exceptions import ConfigFileNotFound, DuplicateAreaName
 from antarest.core.model import JSON
 from antarest.study.business.model.area_model import (
     AreaCreationDTO,
     AreaInfoDTO,
     AreaOutput,
     AreaType,
-    LayerInfoDTO,
     UpdateAreaUi,
 )
 from antarest.study.business.model.area_properties_model import (
@@ -39,11 +38,9 @@ from antarest.study.storage.rawstudy.model.filesystem.config.model import Area, 
 from antarest.study.storage.rawstudy.model.filesystem.config.thermal import parse_thermal_cluster
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.variantstudy.model.command.create_area import CreateArea
-from antarest.study.storage.variantstudy.model.command.icommand import ICommand
 from antarest.study.storage.variantstudy.model.command.remove_area import RemoveArea
 from antarest.study.storage.variantstudy.model.command.update_area_ui import UpdateAreaUI
 from antarest.study.storage.variantstudy.model.command.update_areas_properties import UpdateAreasProperties
-from antarest.study.storage.variantstudy.model.command.update_config import UpdateConfig
 from antarest.study.storage.variantstudy.model.command_context import CommandContext
 
 logger = logging.getLogger(__name__)
@@ -253,155 +250,6 @@ class AreaManager:
         file_study = study.get_files()
         area_ids = list(file_study.config.areas)
         return _get_ui_info_map(file_study, area_ids)
-
-    def get_layers(self, study: StudyInterface) -> List[LayerInfoDTO]:
-        file_study = study.get_files()
-        area_ids = list(file_study.config.areas)
-        ui_info_map = _get_ui_info_map(file_study, area_ids)
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-        if not layers:
-            layers["0"] = "All"
-        return [
-            LayerInfoDTO(
-                id=str(layer),
-                name=layers[str(layer)],
-                areas=[
-                    area
-                    for area in ui_info_map
-                    if str(layer) in _get_area_layers(ui_info_map, area)
-                    # the layer 0 always display all areas
-                    or str(layer) == "0"
-                ],
-            )
-            for layer in layers
-        ]
-
-    def update_layer_areas(self, study: StudyInterface, layer_id: str, areas: List[str]) -> None:
-        logger.info(f"Updating layer {layer_id} with areas {areas}")
-        file_study = study.get_files()
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-        if layer_id not in [str(layer) for layer in list(layers.keys())]:
-            raise LayerNotFound
-        areas_ui = file_study.tree.get(["input", "areas", ",".join(file_study.config.areas), "ui"])
-        # standardizes 'areas_ui' to a dictionary format even if only one area exists.
-        cfg_areas = list(file_study.config.areas)
-        if len(cfg_areas) == 1:
-            areas_ui = {cfg_areas[0]: areas_ui}
-
-        existing_areas = [
-            area for area in areas_ui if "ui" in areas_ui[area] and layer_id in _get_area_layers(areas_ui, area)
-        ]
-        to_remove_areas = [area for area in existing_areas if area not in areas]
-        to_add_areas = [area for area in areas if area not in existing_areas]
-        commands: List[ICommand] = []
-
-        def create_update_commands(area_id: str) -> List[ICommand]:
-            return [
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/layerX",
-                    data=areas_ui[area_id]["layerX"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/layerY",
-                    data=areas_ui[area_id]["layerY"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/ui/layers",
-                    data=areas_ui[area_id]["ui"]["layers"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-            ]
-
-        for area in to_remove_areas:
-            area_to_remove_layers: List[str] = _get_area_layers(areas_ui, area)
-            if layer_id in areas_ui[area]["layerX"]:
-                del areas_ui[area]["layerX"][layer_id]
-            if layer_id in areas_ui[area]["layerY"]:
-                del areas_ui[area]["layerY"][layer_id]
-            if layer_id in area_to_remove_layers:
-                areas_ui[area]["ui"]["layers"] = " ".join(
-                    [area_layer for area_layer in area_to_remove_layers if area_layer != layer_id]
-                )
-            commands.extend(create_update_commands(area))
-        for area in to_add_areas:
-            area_to_add_layers: List[str] = _get_area_layers(areas_ui, area)
-            if layer_id not in areas_ui[area]["layerX"]:
-                areas_ui[area]["layerX"][layer_id] = areas_ui[area]["ui"]["x"]
-            if layer_id not in areas_ui[area]["layerY"]:
-                areas_ui[area]["layerY"][layer_id] = areas_ui[area]["ui"]["y"]
-            if layer_id not in area_to_add_layers:
-                areas_ui[area]["ui"]["layers"] = " ".join(area_to_add_layers + [layer_id])
-            commands.extend(create_update_commands(area))
-
-        study.add_commands(commands)
-
-    def update_layer_name(self, study: StudyInterface, layer_id: str, layer_name: str) -> None:
-        logger.info(f"Updating layer {layer_id} with name {layer_name}")
-        file_study = study.get_files()
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-        if layer_id not in [str(layer) for layer in list(layers.keys())]:
-            raise LayerNotFound
-        command = UpdateConfig(
-            target=f"layers/layers/layers/{layer_id}",
-            data=layer_name,
-            command_context=self._command_context,
-            study_version=study.version,
-        )
-        study.add_commands([command])
-
-    def create_layer(self, study: StudyInterface, layer_name: str) -> str:
-        file_study = study.get_files()
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-        command_context = self._command_context
-        new_id = max((int(layer) for layer in layers), default=0) + 1
-        if new_id == 1:
-            command = UpdateConfig(
-                target="layers/layers/layers",
-                data={"0": "All", "1": layer_name},
-                command_context=command_context,
-                study_version=study.version,
-            )
-        else:
-            command = UpdateConfig(
-                target=f"layers/layers/layers/{new_id}",
-                data=layer_name,
-                command_context=command_context,
-                study_version=study.version,
-            )
-        study.add_commands([command])
-        return str(new_id)
-
-    def remove_layer(self, study: StudyInterface, layer_id: str) -> None:
-        """
-        Remove a layer from a study.
-
-        Raises:
-            LayerNotAllowedToBeDeleted: If the layer ID is "0".
-            LayerNotFound: If the layer ID is not found.
-        """
-        if layer_id == "0":
-            raise LayerNotAllowedToBeDeleted
-
-        file_study = study.get_files()
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-
-        if layer_id not in layers:
-            raise LayerNotFound
-
-        del layers[layer_id]
-
-        command = UpdateConfig(
-            target="layers/layers/layers",
-            data=layers,
-            command_context=self._command_context,
-            study_version=study.version,
-        )
-        study.add_commands([command])
 
     def create_area(self, study: StudyInterface, area_creation_info: AreaCreationDTO) -> AreaInfoDTO:
         file_study = study.get_files()
