+++ tests/integration/test_integration.py
@@ -23,7 +23,7 @@ from starlette.testclient import TestClient
 
 from antarest.core.serde.ini_reader import read_ini
 from antarest.core.serde.ini_writer import write_ini_file
-from antarest.study.business.area_management import LayerInfoDTO
+from antarest.study.business.model.area_model import LayerInfoDTO
 from antarest.study.storage.variantstudy.model.command.common import CommandName
 from tests.integration.assets import ASSETS_DIR
 from tests.integration.utils import wait_for
